Extended Swedish SSN generator
The ssn for swedish locale now takes a gender preference when generating
the national identifier for a person. The ssn generator can also take a
corporate_type number which will generate a corporate identifier. The
personal and corporate numbers are the same format and identified by the
month digits in the date being equal or greater than 20.